Renamed utils to draw_svg
