+++ src/debug/__init__.py
@@ -1,5 +1,5 @@
 from .debug_pipeline import run_pipeline
-from .utils import draw_svg
+from .draw_svg import draw_svg
 
 __all__ = [
     "run_pipeline",
