+++ src/debug/draw_svg.py
